Fixed resnet module
