+++ lucent/modelzoo/resnet/__init__.py
@@ -1,0 +1,1 @@
+from lucent.modelzoo.resnet.resnet import *
