+++ lucent/modelzoo/resnet/resnet.py
@@ -7,6 +7,8 @@ from torchvision.models._api import register_model, Weights, WeightsEnum
 from torchvision.transforms._presets import ImageClassification
 from torchvision.models._meta import _IMAGENET_CATEGORIES
 from torchvision.models._utils import _ovewrite_named_param
+from torchvision.models.resnet import ResNet18_Weights
+
 from typing import Any, Callable, List, Optional, Type, Union
 
 _COMMON_META = {
@@ -356,5 +358,12 @@ def build_resnet(
 
     if weights is not None:
         model.load_state_dict(weights.get_state_dict(progress=progress))
+
+    return model
+
+def resnet18(use_linear_modules=False, skip_batchnorm=True):
+    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+    weights = ResNet18_Weights.verify("IMAGENET1K_V1")
+    model = build_resnet(BasicBlock, [2, 2, 2, 2], weights=weights, progress=True, use_linear_modules_only=use_linear_modules, skip_batchnorm = skip_batchnorm)
 
     return model
